refactor(agent): relocate ResponseCycleFlow and agentCreator out of the flow kernel
Item 9 of docs/proposals/2026-08-14-delegation-flow-substrate-consolidation.md
("Misfiled and mislabeled modules"), executing D6's own fence.

Applied the single-consumer test to every module in core/flows:

- ResponseCycleFlow -> implementations/flows/reflection/ (sole consumer:
  ResponseCycleNode).
- ToolUseRoundFlow + toolUseRound/{roundShared,ToolUseRoundPrepNode,
  ToolUseProcessNode,ToolUseDispatchNode} -> implementations/flows/tooluse/
  (sole consumer: ToolUseCycleNode); this is the same split D6's own
  measurement already used ("tooluse 1,492 + toolUseRound 1,418; shared
  kernel 955").
- toolCallParsing has four model-handler consumers outside both families, so
  it stays in the kernel, moving up one level now that its directory left.
- CycleServices and IToolUseSession stay: ResponseCycleServices and
  ToolUseRoundServices are one shared CycleRunServices base used by both
  families.
- agentCreator -> implementations/agentCreator/ — it uses no flow machinery
  (no BaseNode, no Flow, no transitions) and never belonged under flows/.

No re-export shims: every importer moved to the new path. The core README's
flows row now states the kernel membership rule, and its TaskState row no
longer presents a frozen wire projection (the CLI NDJSON setTaskState
payload) as a live state model.

Ratchets: host-agent-import-baseline renames the extension's agentCreator
specifier (count unchanged — a rename, not a widen); shared-schemas
baseline renames ResponseCycleFlow's path in place (regenerating the whole
file would have swept in unrelated surface drift).
