+++ packages/extension/src/commands/agent/agentCreatorCommands.ts
@@ -9,7 +9,7 @@ import {
   type CreatorConfig,
   TOOL_GROUPS,
   runAgentCreator,
-} from '@agent/implementations/flows/agentCreator/agentCreatorFlow';
+} from '@agent/implementations/agentCreator/agentCreatorFlow';
 import { renderAgentTemplateString } from '@agent/templates/agentTemplateRenderer';
 import { settleQuickInput } from '@commands/_shared/quickInputUtils';
 import { agentDirectories } from '@frontend/agents/AgentDirectoryManager';
