+++ src/agent/implementations/flows/reflection/ResponseCycleFlow.ts
@@ -23,16 +23,15 @@ import {
 import type { ProviderUsage } from '@agent/core/usage/ResponseUsage';
 import { K_SLICE } from '@agent/core/constants';
 import { useLaunchRunContext } from '@agent/runtime/RunContext';
+import { FlowTransition } from '@agent/core/flows/FlowTransitions';
+import { ModelInvocationNode } from '@agent/core/flows/ModelInvocationNode';
+import type { ResponseCycleServices } from '@agent/core/flows/CycleServices';
 import type { ToolDefinition } from '@model/ToolDefinition';
 import { MESSAGE_TYPES, AgentFileLocationSchema } from '@shared/schemas';
 import { OUTPUT_END_TAG } from '@shared/schemas/output';
 import { toErrorMessage } from '@utils/errors/errorMessage';
 import { AbsoluteFS } from '@utils/files/absoluteFS';
 import { extractScratchpad } from '@utils/text/xmlExtraction';
-
-import { FlowTransition } from './FlowTransitions';
-import { ModelInvocationNode } from './ModelInvocationNode';
-import type { ResponseCycleServices } from './CycleServices';
 
 // ============================================================================
 // Cycle Fields Schema (Extends Base)
