+++ src/agent/implementations/flows/tooluse/nodes/ToolUsePrepareNode.ts
@@ -16,7 +16,7 @@ import type { ToolUseRunShared, CyclePrepResult } from './types';
  * message array or restore from a persisted snapshot.
  *
  * This is the outer session-level prep. Compare `ToolUseRoundPrepNode` in
- * `core/flows/toolUseRound/`, which is the inner per-LLM-call prep node that
+ * `../toolUseRound/`, which is the inner per-LLM-call prep node that
  * runs at the start of every model invocation inside `ToolUseRoundFlow`.
  */
 export class ToolUsePrepareNode<C> extends Node<
