+++ src/agent/modelHandlers/openrouter/modelHandlerOpenRouterNative.ts
@@ -4,7 +4,7 @@ import { ModelProvider, ReasoningEffort } from 'llm-zoo';
 
 // Local imports
 import type { StreamHandle } from '@agent/trace';
-import { parseToolInput } from '@agent/core/flows/toolUseRound/toolCallParsing';
+import { parseToolInput } from '@agent/core/flows/toolCallParsing';
 import type { AgentWorkspaceState } from '@agent/core/state/AgentWorkspaceState';
 import type { NormalizedUsage } from '@agent/types/NormalizedUsage';
 import type { MediaEntry } from '@agent/types/mediaTypes';
